requirements/base.txt
Django==1.9.7
djangorestframework==3.3.3
matplotlib==1.5.1
numpy==1.11.1
Pillow==3.3.0
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.1
psycopg2==2.6.2